Wed, Sep 6, 2023, 11:48 AM +02:00

+++ readability/readability.c
@@ -17,8 +17,8 @@ int main(void)
     int words = count_words(text);
     int sentences = count_sentences(text);
 
-    float L = (letters/words) * 100;
-    float S = (sentences/words) * 100;
+    float L = ((float)letters/words) * 100;
+    float S = ((float)sentences/words) * 100;
 
     int index = round((0.0588 * L) - (0.296 * S) - 15.8);
     if (index < 1)
